Fiz a rota findById

+++ project_26/src/main/java/com/gabriel/demo/entities/Order.java
@@ -11,7 +11,7 @@ public class Order {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
     private Instant moment;
 
     @ManyToOne
@@ -21,17 +21,17 @@ public class Order {
     public Order() {
     }
 
-    public Order(Integer id, Instant moment, User client) {
+    public Order(Long id, Instant moment, User client) {
         this.id = id;
         this.moment = moment;
         this.client = client;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
